Add service/package-lock.json to the project
A 'package-lock.json' file has been added to the 'service' directory enhancing the project's dependency management. This file locks the versions of the project's dependencies ensuring that the project will always use the same versions thus enabling consistent behavior across environments. The set versions of dependencies are compiled and specified in the 'package-lock.json' file.