Refactor and improve token handling in ConfigService
The method loadFromVault in ConfigService has undergone an overhaul for better token management. The same previously-read token is now reused instead of being re-read for the jwt field of the request body. On top of that, it ensures token is read from file whenever it exists at the designated secret location. Lastly, a logging implementation has been added when the data is successfully loaded from Vault.