Update .gitignore, Dockerfile, and add preflight script
This commit updates the .gitignore files globally and in the service folder, modifying the Dockerfile to simplify the file permissions configuration and adding a preflight script in service folder. The script retrieves the Kubernetes secrets before running the app, ensuring that the application has the correct permissions and secrets prior to start-up. This will improve the security and efficiency of the deployment process.