Update Node.js version and install additional packages in Dockerfile
This commit updates the Node.js version to its non-minimal variant in the production stage of the Dockerfile. Additionally, it installs the curl and jq packages which are crucial for data manipulation and REST API interactions in UNIX-based systems. These changes aim to provide expanded functionality and better performance for the application's deployment.