Add nonce attribute to scripts and changes to DB_URL usage in preflight.sh
This commit includes the addition of the nonce attribute to scripts in main.njk for better security. Also, instead of throwing an error if DB_URL is not set in preflight.sh, the script now sends a curl request to fetch the database file directly. This caters for situations where the DB_URL might not be set prior to execution.